fix(safety): bmsWatchdogExpired usa uint32_t (wrap de millis correcto en cualquier host)
El test del wrap dependia de la plataforma: 'unsigned long' es 32 bits en
Windows (LLP64) pero 64 en Linux/Mac (LP64), asi que la resta no envolvia igual
y el test daba falso negativo en hosts LP64 (codigo correcto, test enganoso).
Fijar la firma a uint32_t hace que la aritmetica envuelva en 32 bits en target Y
en cualquier host, y de paso el tipo coincide con lo que millis() devuelve en el
STM32. Contrato mas firme, no solo arreglo del test.
Añade ademas test_buzzer_off_on_late_update (el buzzer se apaga al siguiente
update aunque llegue tarde: el corte es por polling, no por tiempo absoluto).

+++ src/SafetyLogic.h
@@ -5,13 +5,20 @@
 // Así se pueden testear en native (host) de forma determinista. El firmware
 // (main.cpp / InverterControl.cpp) las llama; los tests las incluyen directas.
 //==============================================================================
+#include <cstdint>
 
 // ¿Ha expirado el watchdog del BMS? true si han pasado MÁS de 'timeout' ms desde
 // la última trama del BMS (last). Con true, el VCU debe considerar el SDC NO
-// presente (fail-safe: sin tramas, no se fía). La resta en unsigned es robusta
-// ante el wrap de millis() (~49 días).
-inline bool bmsWatchdogExpired(unsigned long last, unsigned long now, unsigned long timeout) {
-    return (now - last) > timeout;
+// presente (fail-safe: sin tramas, no se fía).
+//
+// Tipos uint32_t a propósito: millis() en STM32 devuelve uint32_t, y la resta en
+// 32 bits envuelve correctamente al pasar los ~49 días (el wrap de millis()).
+// Si la firma fuese 'unsigned long', en un host LP64 (Linux/Mac, long=64 bits) la
+// resta NO envolvería como en el target -> el test del wrap daría falso negativo.
+// Con uint32_t el comportamiento es idéntico en target y en cualquier host.
+// El cast del return es defensivo (por si se compila con tipos más anchos).
+inline bool bmsWatchdogExpired(uint32_t last, uint32_t now, uint32_t timeout) {
+    return (uint32_t)(now - last) > timeout;
 }
 
 // Factor [0..1] de corte SUAVE por subtensión del pack, a multiplicar por el
